[clangd] Extract FileSystemProvider into a separate header. NFC
Reviewers: sammccall

Reviewed By: sammccall

Subscribers: mgorny, ioeric, MaskRay, jkorous, cfe-commits

Differential Revision: https://reviews.llvm.org/D49142

git-svn-id: https://llvm.org/svn/llvm-project/clang-tools-extra/trunk@336909 91177308-0d34-0410-b5e6-96231b3b80d8
